一部の独立性の高いルーチンをライブラリとして切り離す。
