+++ calctweet.py
@@ -7,14 +7,14 @@ import requests
 import argparse
 import numpy as np
 import sys
-import os
-import csv
 import unicodedata
 import configparser
-import urllib
-import webbrowser
 from pathlib import Path
 import json
+
+from lib.setting import setting_file_path
+from lib.twitter import create_access_key_secret
+
 
 progname = "FGOツイートスクショチェック"
 version = "0.1.0"
@@ -214,54 +214,6 @@ def calc_diff(report_dic, image_dic, inverse=False):
                     result_dic[item] = report_dic[item] + image_dic[item]                    
     return result_dic
 
-def get_oauth_token(url:str)->str:
-    querys = urllib.parse.urlparse(url).query
-    querys_dict = urllib.parse.parse_qs(querys)
-    return querys_dict["oauth_token"][0]
-
-def create_access_key_secret(CONSUMER_KEY, CONSUMER_SECRET):
-#if __name__ == '__main__':
-
-    auth = tweepy.OAuthHandler(CONSUMER_KEY, CONSUMER_SECRET)
-
-    try:
-        redirect_url = auth.get_authorization_url()
-        print ("次のURLをウェブブラウザで開きます:",redirect_url)
-    except tweepy.TweepError:
-        print( "[エラー] リクエストされたトークンの取得に失敗しました。")
-        sys.exit(1)
-
-
-    oauth_token = get_oauth_token(redirect_url)
-    print("oauth_token:", oauth_token)
-    auth.request_token['oauth_token'] = oauth_token
-
-    # Please confirm at twitter after login.
-    webbrowser.open(redirect_url)
-
-#    verifier = input("You can check Verifier on url parameter. Please input Verifier:")
-    verifier = input("ウェブブラウザに表示されたPINコードを入力してください:")
-    auth.request_token['oauth_token_secret'] = verifier
-
-    try:
-        auth.get_access_token(verifier)
-    except tweepy.TweepError:
-        print('[エラー] アクセストークンの取得に失敗しました。')
-
-    print("access token key:",auth.access_token)
-    print("access token secret:",auth.access_token_secret)
-
-    config = configparser.ConfigParser()
-    section1 = "auth_info"
-    config.add_section(section1)
-    config.set(section1, "ACCESS_TOKEN", auth.access_token)
-    config.set(section1, "ACCESS_SECRET", auth.access_token_secret)
-    config.set(section1, "CONSUMER_KEY", CONSUMER_KEY)
-    config.set(section1, "CONSUMER_SECRET", CONSUMER_SECRET)
-
-    settingfile = os.path.join(os.path.dirname(__file__), 'setting.ini')
-    with open(settingfile, "w") as file:
-        config.write(file)
 
 def meke_output(status, args):
     if 'RT @' not in status.full_text \
@@ -374,7 +326,7 @@ def get_tweet_auto(args, api, last_id):
     
 if __name__ == '__main__':
     last_id = -1
-    settingfile = os.path.join(os.path.dirname(__file__), 'setting.ini')
+    settingfile = setting_file_path()
     config = configparser.ConfigParser()
     try:
         config.read(settingfile)
